added simple role based access

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\PolicyController;
+use App\Http\Middleware\AdminMiddleware;
 
 Route::get('/login', [AuthController::class, 'index'])->name('login');
 Route::post('/login', [AuthController::class, 'login'])->name('login');
@@ -28,5 +29,5 @@ Route::group(['middleware' => 'auth'], function () {
 
     Route::put('/policies/{id}', [PolicyController::class, 'update'])->name('policies.update');
 
-    Route::delete('/policies/{id}', [PolicyController::class, 'destroy'])->name('policies.destroy');
+    Route::delete('/policies/{id}', [PolicyController::class, 'destroy'])->name('policies.destroy')->middleware(AdminMiddleware::class);
 });
